+++ test/test_network.h
@@ -119,7 +119,7 @@ TEST(network, add) {
   network<sequential> net;
   net << convolutional_layer(32, 32, 5, 3, 6, padding::same);
 
-  EXPECT_EQ(net.depth(), static_cast<size_t>(1));
+  EXPECT_EQ(net.depth(), 1);
 }
 
 TEST(network, manual_init) {
@@ -134,9 +134,9 @@ TEST(network, manual_init) {
   vec_t *c1_b = net[0]->weights()[1];
   vec_t *f1_w = net[1]->weights()[0];
 
-  EXPECT_EQ(c1_w->size(), static_cast<size_t>(9));
-  EXPECT_EQ(c1_b->size(), static_cast<size_t>(1));
-  EXPECT_EQ(f1_w->size(), static_cast<size_t>(2));
+  EXPECT_EQ(c1_w->size(), 9);
+  EXPECT_EQ(c1_b->size(), 1);
+  EXPECT_EQ(f1_w->size(), 2);
 
   *c1_w = {0, 1, 2, 3, 4, 5, 6, 7, 8};
   *c1_b = {1};
