+++ tiny_dnn/core/kernels/tiny_deconv2d_kernel.h
@@ -31,12 +31,12 @@ inline void tiny_deconv2d_kernel(const deconv_params &params,
         const float_t *pw = &W[idx];
 
         idx = params.in.get_index(0, 0, inc);
-        assert(static_cast<size_t>(sample) < in.size() &&
+        assert(sample < in.size() &&
                idx <= in[sample].size());
         const float_t *pi = &in[sample][idx];
 
         idx = params.out.get_index(0, 0, o);
-        assert(static_cast<size_t>(sample) < out.size() &&
+        assert(sample < out.size() &&
                idx <= out[sample].size());
         float_t *pout = &out[sample][idx];
